Initial changes for XASTO and LXASTO
Since I'm running out of opcodes in the HP-42S parameterized extension
space, I'm defining extended extensions, using the opcode 0xa7 followed
by a secondary opcode byte in the range 0x01-0x7f. It's a bit of a hack
but it works and doesn't require lots of code changes.
This provides the six slots for XASTO and LXASTO (each in suffix
(register or IND register), alpha, and IND alpha varieties), which is
where the whole Long String thing will start.