Adding rttypes field to command_spec
I wanted to do this in a way that wouldn't cause the record to grow, so
I shuffled things around a bit and merged the two flag bytes, since one
of them used only five bits and the other only two. The rttypes field
will be used to indicate which types a function will accept, so that
handle() can perform a preliminary argument check. This will add a
slight overhead, but has the advantage of removing a bunch of parameter
check boilerplate, and, more importantly, avoids the need to *add* a
ton of boilerplate for the list type, when it becomes a full citizen of
user space. Not to mention other types that may be added later, like
units, programs, and directories.