Populated rttypes in cmd_array
Up next: remove the now-redundant type checks wherever they are now, in
fact, redundant. Check the seq? comments for cases where the handle()
type check would be premature. Check all the cases where we have ALLT
or FUNC but where types other than the original five may not be handled
correctly (missing "everything else" -> ERR_INVALID_TYPE fallback). And
handle SUBSTR and POS, if there is an elegant way to do so. Those are
probably not functions that are going to show up in expressions, but
still...