Implemented HEAD, APPEND, and EXTEND.
APPEND and EXTEND replace CONCAT. They do the same thing, except APPEND
adds a list to a list as a single element, while EXTEND adds a list to
a list element-by-element. The nomenclature was gleaned from Python.

Note: Don't trust APPEND and EXTEND with strings yet, the long string
patch is not yet in place, and the code as is has the potential to
corrupt memory.