is_string[] cosmetic fix
This may look excessively anal, but with the big strings, the is_string
array can contain three values, not two, so it's a matter of good style
to do proper comparisons rather than treat it as boolean. The compiler
will generate the same code anyway, but humans will have an easier time
grokking it.