Removing the strictness requirement from alias lookups
In addition to the function aliases defined on the HP-42S (CHS for +/-,
OCT for →OCT, etc.), Free42 defined additional aliases, like STO/ for
STO÷ and X#0? for X≠Y?, etc., to help with pasting program listings. So
far, these added aliases were only recognized by Paste and for Direct
Command Mapping, but not in the CUSTOM menu or when spelling out a
function with XEQ. I'm leaving it like that in Free42, but allowing all
aliases for CUSTOM and XEQ in Plus42 now.