Looks like I fixed the direct solution crash
All caused by forgetting to duplicate a vartype before handing it to
store_var(). Doh.
Committing with all the debug junk still in place; I'll remove those in
the next commit, but this way I'll be able to retrieve them easily,
just in case I ever need to look at the system's guts this way again.