Fixed a misguided fix in equation editor
Clearing eqns was wrong: that variable is just an alias to the EQNS
user-code variable.
EQNS should be managed more safely, i.e. by re-getting it (and updating
num_eqns!) whenever it is needed, and never holding on to it in a
persistent manner. If num_eqns goes out of range because of deletions
behind the equation editor's back, the editor should just pad out the
array (with what?) to make things fit, and of course always disentangle
eqns before making modifications to it.