Emergency hack to deal with equation leak
Of course this isn't a proper fix; that would be to identify the source
of the leak, or to make sure the previous fixes have actually taken
care of it already, and what we've seen lately was just leftover
garbage faithfully preserved in the state file.  This fix compares
refcounts on equation_data records to the number of references actually
found in the state file; records with no living references are deleted,
and all others have their refcounts reset to what was found in the
file. At the very least, this will prevent leaked equations from piling
up without limit.
Also, if a file named $HOME/plus42-memory-stats.txt exists, the
cleaner-upper will write a detailed report of what it finds and what
actions it performs, appending the report to the end of the file.