More complex conversions; more functions in menus
Added Σ and ∫ to the PGM.FCN menu; added RCOMPLX, PCOMPLX, and INT to
the CONVERT menu; added hyperbolics, LN1+x and E^X-1 to the TOP.FCN
menu; moved some other functions around to what will hopefully be their
final locations.
ANGLE, RADIUS, XCOORD, and YCOORD can now be used with a single
(complex) argument.
Added REAL?, CPX?, MAT? and LIST?
Renamed the parameterized EVAL to EVALN, and restored the parameter-
less EVAL. Note, however, that the new EVAL is not the same as the old
EVAL; it behaves like EVALN ST X, meaning, it acts like a recall, not
like a unary operator.